Time: 96 ms (8.39%), Space: 27.5 MB (97.90%) - LeetHub

+++ 1354-construct-target-array-with-multiple-sums/1354-construct-target-array-with-multiple-sums.cpp
@@ -1,15 +1,15 @@
 class Solution {
 public:
     bool isPossible(vector<int>& target) {
-        long sum = 0;
-        long maxi = 0;
+        long long int sum = 0;
+        long long int maxi = 0;
         
         for(int i=0;i<target.size();i++){
             sum += target[i];
         }
         maxi = max_element(target.begin(),target.end()) - target.begin();
         
-        long diffr = sum - target[maxi];
+        long long int diffr = sum - target[maxi];
         
         if(target[maxi] == 1 || diffr == 1) return true;
         if(diffr > target[maxi] || diffr == 0 || target[maxi]%diffr == 0) return false;
